+++ test/helpers/fixture.js
@@ -234,4 +234,40 @@ export default class Fixture {
       ],
     };
   }
+
+  static eventBody() {
+    const data = {
+      result: {
+        id: 'batch_123...',
+        object: 'Batch',
+        mode: 'test',
+        state: 'created',
+        num_shipments: 0,
+        reference: null,
+        created_at: '2022-07-26T17:22:32Z',
+        updated_at: '2022-07-26T17:22:32Z',
+        scan_form: null,
+        shipments: [],
+        status: {
+          created: 0,
+          queued_for_purchase: 0,
+          creation_failed: 0,
+          postage_purchased: 0,
+          postage_purchase_failed: 0,
+        },
+        pickup: null,
+        label_url: null,
+      },
+      description: 'batch.created',
+      mode: 'test',
+      previous_attributes: null,
+      completed_urls: null,
+      user_id: 'user_123...',
+      status: 'pending',
+      object: 'Event',
+      id: 'evt_123...',
+    };
+
+    return Buffer.from(JSON.stringify(data), 'utf8');
+  }
 }